Add ActionButton widget
