+++ lib/main.dart
@@ -3,8 +3,8 @@ import 'package:flutter/material.dart';
 import 'package:places/ui/res/themes.dart';
 
 // import 'package:places/ui/screens/visiting_screen.dart';
-// import 'package:places/ui/screens/sight_details.dart';
-import 'package:places/ui/screens/sight_list_screen.dart';
+import 'package:places/ui/screens/sight_details.dart';
+// import 'package:places/ui/screens/sight_list_screen.dart';
 
 void main() {
   runApp(App());
@@ -26,7 +26,7 @@ class _AppState extends State<App> {
       darkTheme: darkTheme,
       themeMode: isDarkMode ? ThemeMode.dark : ThemeMode.light,
       title: "My first App",
-      home: SightListScreen(changeThemeMode: changeThemeMode),
+      home: SightDetails(changeThemeMode: changeThemeMode),
     );
   }
